Casino Scene + Drawable Cards
I finished the background for the Casino scene. It has the green Killer table with three chairs, and the Casino is in the background behind it.

I also loaded all of the card skins into an array in the Card class and am able to create new Card objects easily from the Killer.java class.

Next time, I think I will try to implement some sort of "selected" feature for the Card class so that when a Card has been selected it is outlined with a brighter color. I could also try to make it so that the cards appear a different color when the mouse hovers over them.

Eventually, I'll also need to make some sort of display for the center of the table so that the user can see what cards the opponents have played. Additionally, I should add some feature that will show the user the backs of the opponents cards so that they can see how many cards the opponents have left.