Switches to Organizational Style
I'm switching things up a bit so that Button and Scene are abstract classes.

This way, the actionPerformed for a Button must be defined each time an attempt is made to create a Button.

And for a Scene, I would plan to make separate subclasses for Scenes where each subclass is a different Scene. Title Screen and Game Table will really be the only two, so this won't result in excessive numbers of class files being created and it will allow for me to define specific load( ), update( ) and draw( ) functions.

TitleScreen and CasinoTable are the two new Scene subclasses. CasinoTable will at some point need to be able to draw various cards at different locations and in different positions/styles, and I think I'll eventually store those cards in the GameHandler.java file.