Deck Class, Texture Updates, Buttons, & New Images
Created the Deck class which handles the grouping of Cards. This class builds off of and utilizes the Card class, and I plan for the GameHandler class to build off of the Deck class. The constructor currently takes no arguments and makes a standard full 52 Card deck. There is a shuffle method to randomize Card order, a sort method to organize Cards from lowest to highest number value, a deal method to deal out a specific number of Cards to a specific number of hands, and a draw method to draw all of the Cards in the deck.

Modified Card.java so that it could import textures with less redundant code (it now will import them once on the first call to the Card constructor, and all the skins will be stored in the static skins 2D array in the Card class).

Modified Button so that it has a Texture array containing different skins (which are really just different sizes) for the Button. The constructor also takes a buttonType argument (int) that inidcates what type of Button to make. There are two final static variables, LONG and SHORT, which store integer values that correlate to an index in the skins array. Additionally, Button no longer takes scaleX and scaleY, as there is no scaling for the Buttons since they have specific skins for specific sizes.

Cleaned up the Button constructor calls in CasinoTable and TitleScreen so that the x and y coordinates are now specific numbers rather than equations (better for readability).

Created new images for a shortButton and for the backs of cards (distorted perspectives angled right and left as well as a straight one)