GameHandler, Card.java Changes, & Code Tidying
Started working on the GameHandler class which, you guessed it, handles the game! As of now, it is able to set up a Deck of 52 cards, deal those cards out into four different hands, and then draw each of those hands on the screen.

I made some slight changes to Card.java so that the skins for the back of the card were easier to access and use. I also implemented some more getter and setter methods. Additionally, I've made it so that when a mouse is hovering over the Card it'll darken in color to indicate hovering. And I also made sure that you can only select your own cards (and only darken your own cards when you hover).

Every single file I have for this project has now been fully updated and organized with comments. Everything looks very clean and I'm having a very easy time navigating around. This project and the programming for games pokemon project have helped me develop a coding organization style that works really great for me.