Card Movement
I started working on Card movement. Currently when a Card is clicked on and selected it will rise slightly above the non-selected cards. This provides even further clarity for knowing which cards you have selected.

Next I plan on implementing the movement for Cards to be played and laid down in the middle of the table. This will require the Cards that are being played to be resized and flipped, and I can do both in code without needing to make new Cards (hopefully).