GameHandler Work
I'm working on the visuals of the game right now, like how the Cards will move smoothly when played and the hand will readjust and center as Cards are being removed.

There are currently a lot of bugs with this, but I'm taking a break and will come back to this later.