Card Movement and Visuals
I implemented more card movement and visual work. When the cards are played, they shrink in size as they move to the center of the screen. When a new turn begins and more cards are played, any cards that were in the center of the table are erased and new ones are drawn in their place.