Implemented Turn System
I implemented a rotating turn system for the game. When launched, the game will determine which hand has the three of spades and will set that hand to have the first turn. After the Cards are played, the turn is shifted to the left (as is done in real Killer).

Currently there are some bugs that I am working on. I need to ensure that the Cards for the bot hands are flipped as they are played. Additionally, I need to figure out how to make the bots pick the best move. First I'm going to work on the logic for checking a valid move, and then after I have that code it might be easier to make the bot logic.