Implemented Rounds and Card Value Checking
I made a few functions that could handle the changing and starting of rounds in Killer. I also made a roundType field that keeps track of whether its singles, doubles, triples, or consecutives on the table for the current round.

Currently there is a bug on line 232 of the GameHandler where one of the for loops I wrote is quite simply just not running, very strange error. I don't have time to look now, but next time I will figure that out.

I also made functions the validate a move of selected cards. There is a function for each round type, singles, doubles, triple, and consecutives. Each function is capable of determining whether or not the cards are valid based on the roundType, the cards that are already on the table (if any), and of course the cards that are to be played.