Almost Fully Functional
Made a lot of progress today. I got the turn sequence system to work, so turns will now cycle through the four hands. I also incorporated logic for the bots so that they can choose their own cards.

It isn't the most complicated AI I could have come up with, in fact, I had other more fleshed out ideas for how to code the bot behavior. Ultimately though, I ended up coding them to always play their cards if they're able to and to always play the lowest cards they possibly can. Even though this is a simple strategy, it has actually proven to be quite effective in testing. Since the bots are capable of winning rounds and starting off rounds with their own set of cards, it can make things go in their favor a lot more.

There are a few bugs I still need to work on. Currently, the bots will sometimes pass their turn at the start of a round. Additionally, I can see the cards of opposing players flash red as they are selected which I shouldn't (that needs to be fixed by modifying Card.java). I also need to make it so that raised cards in the player view go back down after trying to be played if they can't be (add an error noise as they fall back to their normal position)

Overall, my code has remained quite organized with the exception of the GameHandler.java file. The code in there is well-documented and fully explained with comments, but it is disorganized in the sense that there is lots of redundancy. I was just trying to get things to work, so I wasn't necessarily thinking of the best way to make them work. It looks a bit sloppy and I'm hoping to polish that file up a bit. I might split the bot functionality into its own file since that seems to be making up a large amount of the code in the GameHandler.java file.

My plan is to get the game ready for full debugging by the end of tomorrow, meaning that all functionality will be there and that all I'll need to do is play the game and test for bugs. I'll do that bug testing on Tuesday.