Created Separate Bot File & Bug Fixes
Created the Bot's playing behavior code in the new Bot.java class. The way I've programmed them to behave is to consider the round type,  consider all valid moves based on the round type, and select a random valid move out of all those that were considered. It's a pretty effective way to program them because it is good for a beginner's level introduction to the game. The bots obviously aren't considering their cards, which might make things seem a bit easy, but in reality a beginner player also doesn't really understand the full impact of their cards and the moves they play right off the bat so I think it balances out. I still need to add the logic for the Bot to deal with consecutives, and there are a couple bugs I need to look into that may or may not be caused by the Bot code. For some reason they can pass on their first turn sometimes, and additionally there are times where an invalid selection is allowed to be thrown by a bot. Very odd.

I also made some bug fixes. Bot cards no longer appear highlighted before playing and the win screen and lose screen are in and they work now. Another thing I need to implement is an error noise for when the player tries to play invalid cards, and when this occurs I also need all of the player's selected cards to lower back down to the rest of the cards.