Nearing Completion
The Bot behavior is looking much better than before. Singles, triples, and consecutives are working very nicely, but there is a slight issue with doubles. For some reason, bots are able to get through the doubles validation functions even though the Cards they selected don't have the same value. Odd.

I need to figure out why the Buttons dont darken the same color after one is clicked on when its not the Player's turn After I figure out those issues, I just need to make a rules sheet and have it be displayable in game (when the button is pressed, the rules sheet image can be drawn over the screen