Bug Fix
Fixed a bug that allowed the Player to throw more Cards than they should be able to based on the round type. For example, on a singles round the Player was able to throw doubles when they shouldn't have been able to. That is now fixed for all round types