Game Seems To Be Fully Debugged + Adding Rules
I had some people play the game and try to break it. One person was able to break it pretty bad and helped me fix a huge bug which was super helpful, and after I fixed that everything seemed to be good from there on out.

I'm adding the final touch to the game, which will be a rules menu in the CasinoTable scene. I created four pretty brief visuals to help explain the game, and I think they do the trick pretty well