Game is Complete
The End is here! Fingers crossed there aren't any bugs I missed. I had several people test it and everything seemed to work well. This was a fun project to work on, and I hope to go back and add some further functionality to the game on my own time.